requirement.txt
flask>=2.0.0
langchain>=0.0.267
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
pypdf>=3.15.1
chromadb>=0.4.13
sentence-transformers>=2.2.2
huggingface-hub>=0.16.4